fix codegen
