+++ rbatis-codegen/src/codegen/syntax_tree_pysql/bind_node.rs
@@ -1,4 +1,4 @@
-use crate::codegen::syntax_tree_pysql::{DefaultName, Name};
+use crate::codegen::syntax_tree_pysql::{DefaultName, Name, ToHtml};
 
 /// Represents a `bind` or `let` node in py_sql.
 /// It's used to assign a value to a variable within the SQL query.
@@ -25,5 +25,12 @@ impl DefaultName for BindNode {
 impl Name for BindNode {
     fn name() -> &'static str {
         "bind"
+    }
+}
+
+
+impl ToHtml for BindNode {
+    fn as_html(&self) -> String {
+        format!("<bind name=\"{}\" value=\"{}\"/>", self.name, self.value)
     }
 }
