+++ rbatis-codegen/src/codegen/syntax_tree_pysql/choose_node.rs
@@ -1,4 +1,4 @@
-use crate::codegen::syntax_tree_pysql::{Name, NodeType};
+use crate::codegen::syntax_tree_pysql::{Name, NodeType, ToHtml};
 
 /// Represents a `choose` node in py_sql.
 /// It provides a way to conditionally execute different blocks of SQL, similar to a switch statement.
@@ -25,5 +25,23 @@ pub struct ChooseNode {
 impl Name for ChooseNode {
     fn name() -> &'static str {
         "choose"
+    }
+}
+
+
+impl ToHtml for ChooseNode {
+    fn as_html(&self) -> String {
+        let mut childs = String::new();
+        for x in &self.when_nodes {
+            childs.push_str(&x.as_html());
+        }
+        let mut other_html = String::new();
+        match &self.otherwise_node {
+            None => {}
+            Some(v) => {
+                other_html = v.as_html();
+            }
+        }
+        format!("<choose>{}{}</choose>", childs, other_html)
     }
 }
