+++ rbatis-codegen/src/codegen/syntax_tree_pysql/if_node.rs
@@ -1,4 +1,4 @@
-use crate::codegen::syntax_tree_pysql::{Name, NodeType};
+use crate::codegen::syntax_tree_pysql::{Name, NodeType, ToHtml};
 
 /// Represents an `if` conditional node in py_sql.
 /// It executes the nested SQL block if the `test` condition evaluates to true.
@@ -23,5 +23,16 @@ pub struct IfNode {
 impl Name for IfNode {
     fn name() -> &'static str {
         "if"
+    }
+}
+
+
+impl ToHtml for IfNode {
+    fn as_html(&self) -> String {
+        let mut childs = String::new();
+        for x in &self.childs {
+            childs.push_str(&x.as_html());
+        }
+        format!("<if test=\"{}\">{}</if>", self.test, childs)
     }
 }
