+++ rbatis-codegen/src/codegen/syntax_tree_pysql/mod.rs
@@ -13,6 +13,7 @@ pub mod string_node;
 pub mod trim_node;
 pub mod when_node;
 pub mod where_node;
+pub mod to_html;
 
 use crate::codegen::syntax_tree_pysql::bind_node::BindNode;
 use crate::codegen::syntax_tree_pysql::break_node::BreakNode;
@@ -152,175 +153,6 @@ pub trait DefaultName {
 }
 
 /// Convert syntax tree to HTML deconstruction
-pub trait AsHtml {
+pub trait ToHtml {
     fn as_html(&self) -> String;
-}
-
-impl AsHtml for StringNode {
-    fn as_html(&self) -> String {
-        if self.value.starts_with("`") && self.value.ends_with("`") {
-            self.value.to_string()
-        } else {
-            let mut v = self.value.clone();
-            v.insert(0, '`');
-            v.push('`');
-            v
-        }
-    }
-}
-
-impl AsHtml for IfNode {
-    fn as_html(&self) -> String {
-        let mut childs = String::new();
-        for x in &self.childs {
-            childs.push_str(&x.as_html());
-        }
-        format!("<if test=\"{}\">{}</if>", self.test, childs)
-    }
-}
-
-impl AsHtml for TrimNode {
-    fn as_html(&self) -> String {
-        let mut childs = String::new();
-        for x in &self.childs {
-            childs.push_str(&x.as_html());
-        }
-        format!(
-            "<trim prefixOverrides=\"{}\" suffixOverrides=\"{}\">{}</trim>",
-            self.start, self.end, childs
-        )
-    }
-}
-
-impl AsHtml for ForEachNode {
-    fn as_html(&self) -> String {
-        let mut childs = String::new();
-        for x in &self.childs {
-            childs.push_str(&x.as_html());
-        }
-        format!(
-            "<foreach collection=\"{}\" index=\"{}\" item=\"{}\" >{}</foreach>",
-            self.collection, self.index, self.item, childs
-        )
-    }
-}
-
-impl AsHtml for ChooseNode {
-    fn as_html(&self) -> String {
-        let mut childs = String::new();
-        for x in &self.when_nodes {
-            childs.push_str(&x.as_html());
-        }
-        let mut other_html = String::new();
-        match &self.otherwise_node {
-            None => {}
-            Some(v) => {
-                other_html = v.as_html();
-            }
-        }
-        format!("<choose>{}{}</choose>", childs, other_html)
-    }
-}
-
-impl AsHtml for OtherwiseNode {
-    fn as_html(&self) -> String {
-        let mut childs = String::new();
-        for x in &self.childs {
-            childs.push_str(&x.as_html());
-        }
-        format!("<otherwise>{}</otherwise>", childs)
-    }
-}
-
-impl AsHtml for WhenNode {
-    fn as_html(&self) -> String {
-        let mut childs = String::new();
-        for x in &self.childs {
-            childs.push_str(&x.as_html());
-        }
-        format!("<when test=\"{}\">{}</when>", self.test, childs)
-    }
-}
-
-impl AsHtml for BindNode {
-    fn as_html(&self) -> String {
-        format!("<bind name=\"{}\" value=\"{}\"/>", self.name, self.value)
-    }
-}
-
-impl AsHtml for SetNode {
-    fn as_html(&self) -> String {
-        let mut childs_html = String::new();
-        for x in &self.childs {
-            childs_html.push_str(&x.as_html());
-        }
-
-        let mut attrs_string = String::new();
-        if !self.collection.is_empty() {
-            attrs_string.push_str(&format!(" collection=\"{}\"", self.collection));
-        }
-        if !self.skips.is_empty() {
-            attrs_string.push_str(&format!(" skips=\"{}\"", self.skips));
-        }
-        if self.skip_null {
-            attrs_string.push_str(" skip_null=\"true\"");
-        }
-
-        format!("<set{}>{}</set>", attrs_string, childs_html)
-    }
-}
-
-impl AsHtml for WhereNode {
-    fn as_html(&self) -> String {
-        let mut childs = String::new();
-        for x in &self.childs {
-            childs.push_str(&x.as_html());
-        }
-        format!("<where>{}</where>", childs)
-    }
-}
-
-impl AsHtml for NodeType {
-    fn as_html(&self) -> String {
-        match self {
-            NodeType::NString(n) => n.as_html(),
-            NodeType::NIf(n) => n.as_html(),
-            NodeType::NTrim(n) => n.as_html(),
-            NodeType::NForEach(n) => n.as_html(),
-            NodeType::NChoose(n) => n.as_html(),
-            NodeType::NOtherwise(n) => n.as_html(),
-            NodeType::NWhen(n) => n.as_html(),
-            NodeType::NBind(n) => n.as_html(),
-            NodeType::NSet(n) => n.as_html(),
-            NodeType::NWhere(n) => n.as_html(),
-            NodeType::NContinue(n) => n.as_html(),
-            NodeType::NBreak(n) => n.as_html(),
-            NodeType::NSql(n) => n.as_html(),
-        }
-    }
-}
-
-impl AsHtml for Vec<NodeType> {
-    fn as_html(&self) -> String {
-        let mut htmls = String::new();
-        for x in self {
-            htmls.push_str(&x.as_html());
-        }
-        htmls
-    }
-}
-
-pub fn to_html(args: &Vec<NodeType>, is_select: bool, fn_name: &str) -> String {
-    let htmls = args.as_html();
-    if is_select {
-        format!(
-            "<mapper><select id=\"{}\">{}</select></mapper>",
-            fn_name, htmls
-        )
-    } else {
-        format!(
-            "<mapper><update id=\"{}\">{}</update></mapper>",
-            fn_name, htmls
-        )
-    }
 }
