+++ rbatis-codegen/src/codegen/syntax_tree_pysql/otherwise_node.rs
@@ -1,4 +1,4 @@
-use crate::codegen::syntax_tree_pysql::{DefaultName, Name, NodeType};
+use crate::codegen::syntax_tree_pysql::{DefaultName, Name, NodeType, ToHtml};
 
 /// Represents an `otherwise` node in py_sql.
 /// It's used within a `choose` block to provide a default SQL block to execute if none of the `when` conditions are met.
@@ -28,5 +28,15 @@ impl Name for OtherwiseNode {
 impl DefaultName for OtherwiseNode {
     fn default_name() -> &'static str {
         "_"
+    }
+}
+
+impl ToHtml for OtherwiseNode {
+    fn as_html(&self) -> String {
+        let mut childs = String::new();
+        for x in &self.childs {
+            childs.push_str(&x.as_html());
+        }
+        format!("<otherwise>{}</otherwise>", childs)
     }
 }
