+++ rbatis-codegen/src/codegen/syntax_tree_pysql/trim_node.rs
@@ -1,4 +1,4 @@
-use crate::codegen::syntax_tree_pysql::{Name, NodeType};
+use crate::codegen::syntax_tree_pysql::{Name, NodeType, ToHtml};
 
 /// Represents a `trim` node in py_sql.
 /// It's used to remove leading and/or trailing characters from a string.
@@ -21,5 +21,18 @@ pub struct TrimNode {
 impl Name for TrimNode {
     fn name() -> &'static str {
         "trim"
+    }
+}
+
+impl ToHtml for TrimNode {
+    fn as_html(&self) -> String {
+        let mut childs = String::new();
+        for x in &self.childs {
+            childs.push_str(&x.as_html());
+        }
+        format!(
+            "<trim prefixOverrides=\"{}\" suffixOverrides=\"{}\">{}</trim>",
+            self.start, self.end, childs
+        )
     }
 }
